+++ feature/books/domain/src/commonMain/kotlin/com/retro99/books/domain/model/SeriesDomainModel.kt
@@ -4,7 +4,7 @@ data class SeriesDomainModel(
     val uuid: String,
     val name: String,
     val featured: Int?,
-    val position: Int?,
+    val position: Double?,
     val createdAt: String?,
     val updatedAt: String?,
 )
